Extract SaveState into a separate model and service
