--- src/app/Domain/Autosave/saveState.ts
@@ -1,7 +1,0 @@
-import { BusinessObject } from '../Common/businessObject';
-
-export interface SaveState {
-  title: string;
-  description: string;
-  domainStory: BusinessObject[];
-}
